minor2

+++ NeuralNetwork/learning.c
@@ -251,11 +251,11 @@ int main(int argc, char **argv[])
         printf("Usage : \nmain double v  size_t itteration  size_t hidenNumber  size_t lowerBound  size_t upperBound");
         return 1;
     }
-    double v = argv[1];
-    size_t itteration = argv[2];
-    size_t hidenNumber = argv[3];
-    size_t testLen = argv[4];
-    size_t lowerBound = argv[5];
+    double v = (double)(argv[1]);
+    size_t itteration = (size_t)(argv[2]);
+    size_t hidenNumber = (size_t)(argv[3]);
+    size_t testLen = (size_t)(argv[4]);
+    size_t lowerBound = (size_t)(argv[5]);
